implementation/frontend/test-data/requirements.txt
# Mock Python requirements with quantum-vulnerable dependencies
# For testing QuantumGuard CryptoDeps scanner

# Quantum-Vulnerable Packages
pycryptodome==3.19.0
cryptography==41.0.5
rsa==4.9
ecdsa==0.18.0
pyOpenSSL==23.2.0

# Safe dependencies
flask==3.0.0
sqlalchemy==2.0.23
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
pydantic==2.5.2
python-dotenv==1.0.0
boto3==1.29.6
pandas==2.1.3

# Hash usage (info-level, not inherently vulnerable)
hashlib
